Remove database calls from CSVController and handle large files
The code responsible for inserting the transactions extracted from the CSV file is now in TransactionBuilder. TransactionBuilder also now uses fgets and saves the transaction as soon as it extracts it. These two changes prevent memory overload and make the method theoratically able to handle large files. Also notice the use of firstOrCreate which sligtly simplifies the check for an already existing transaction with the same hash when inserting a transaction.