Changed things
Added a date picker so that the user has to pick dates, as opposed to using the pre-defined ones at the time.
Changed the V bar graph so all labels appear at the bottom.
Added a conditionally changing legend so it shows one thing for V and H bar graphs, and something else for Pie graphs.